Bug/Fix Tao don hang YCGC chi tiet

+++ Service.Manafacturing.Model/Outsource/Order/OutsourceOrderDetailModel.cs
@@ -10,7 +10,7 @@ namespace VErp.Services.Manafacturing.Model.Outsource.Order
     {
         public long OutsourceOrderDetailId { get; set; }
         public long OutsoureOrderId { get; set; }
-        public int ObjectId { get; set; }
+        public long ObjectId { get; set; }
         public decimal Price { get; set; }
         public decimal Tax { get; set; }
     }
